+++ waspc/data/Generator/templates/server/src/auth/providers/oauth/handler.ts
@@ -17,10 +17,7 @@ import {
   handleOAuthErrorAndGetRedirectUri,
 } from '../oauth/user.js'
 import { callbackPath, loginPath } from './redirect.js'
-import {
-  onBeforeOAuthRedirectHook,
-  onAfterOAuthTokenReceivedHook,
-} from '../../hooks.js'
+import { onBeforeOAuthRedirectHook } from '../../hooks.js'
 
 export function createOAuthProviderRouter<ST extends StateType>({
   provider,
@@ -118,13 +115,6 @@ function createOAuthCallbackHandler<ST extends StateType>(
     try {
       const oAuthState = validateAndGetOAuthState(stateTypes, provider, req)
       const { accessToken } = await getProviderTokens(oAuthState)
-      if (onAfterOAuthTokenReceivedHook) {
-        // TODO: send state as well so it can be connected to the redirect
-        await onAfterOAuthTokenReceivedHook({
-          req,
-          accessToken,
-        })
-      }
 
       const { providerProfile, providerUserId } = await getProviderInfo({
         accessToken,