+++ waspc/examples/todoApp/src/auth/hooks.ts
@@ -1,5 +1,4 @@
 import type {
-  OnAfterOAuthTokenReceivedHookFn,
   OnAfterSignupHookFn,
   OnBeforeOAuthRedirectHookFn,
   OnBeforeSignupHookFn,
@@ -18,16 +17,10 @@ export const onAfterSignup: OnAfterSignupHookFn = async (args) => {
 }
 
 export const onBeforeOAuthRedirect: OnBeforeOAuthRedirectHookFn = async (
-  args,
+  args
 ) => {
   console.log('req.query before redirect', args.req.query)
   args.url.searchParams.set('someState', '123')
   console.log('redirect to', args.url.toString())
   return { url: args.url }
 }
-
-export const onAfterOAuthTokenReceived: OnAfterOAuthTokenReceivedHookFn =
-  async (args) => {
-    console.log('req.query after callback', args.req.query)
-    console.log('access token', args.accessToken)
-  }