+++ waspc/test/AppSpec/ValidTest.hs
@@ -118,8 +118,7 @@ spec_AppSpecValid = do
                 AS.Auth.onAuthSucceededRedirectTo = Nothing,
                 AS.Auth.onBeforeSignup = Nothing,
                 AS.Auth.onAfterSignup = Nothing,
-                AS.Auth.onBeforeOAuthRedirect = Nothing,
-                AS.Auth.onAfterOAuthTokenReceived = Nothing
+                AS.Auth.onBeforeOAuthRedirect = Nothing
               }
 
       describe "should validate that when a page has authRequired, app.auth is also set." $ do
@@ -165,8 +164,7 @@ spec_AppSpecValid = do
                                     AS.Auth.onAuthSucceededRedirectTo = Nothing,
                                     AS.Auth.onBeforeSignup = Nothing,
                                     AS.Auth.onAfterSignup = Nothing,
-                                    AS.Auth.onBeforeOAuthRedirect = Nothing,
-                                    AS.Auth.onAfterOAuthTokenReceived = Nothing
+                                    AS.Auth.onBeforeOAuthRedirect = Nothing
                                   },
                             AS.App.emailSender =
                               Just
@@ -306,8 +304,7 @@ spec_AppSpecValid = do
                                     AS.Auth.onAuthSucceededRedirectTo = Nothing,
                                     AS.Auth.onBeforeSignup = Nothing,
                                     AS.Auth.onAfterSignup = Nothing,
-                                    AS.Auth.onBeforeOAuthRedirect = Nothing,
-                                    AS.Auth.onAfterOAuthTokenReceived = Nothing
+                                    AS.Auth.onBeforeOAuthRedirect = Nothing
                                   },
                             AS.App.emailSender = emailSender
                           },