+++ waspc/ChangeLog.md
@@ -6,7 +6,7 @@
 
 - Simplified Auth User API: Introduced a simpler API for accessing user auth fields (for example `username`, `email`, `isEmailVerified`) directly on the `user` object, eliminating the need for helper functions.
 - Improved API for calling Operations (Queries and Actions) directly.
-- Auth Hooks: you can now hook into the auth process with `onBeforeSignup`, `onAfterSignup` hooks. You can also modify the OAuth redirect URL with `onBeforeOAuthRedirect` hook and get the provider token with `onAfterOAuthTokenReceived` hook.
+- Auth Hooks: you can now hook into the auth process with `onBeforeSignup`, `onAfterSignup` hooks. You can also modify the OAuth redirect URL with `onBeforeOAuthRedirect` hook.
 
   ```wasp
   app myApp {
@@ -15,7 +15,6 @@
       onBeforeSignup: import { onBeforeSignup } from "...",
       onAfterSignup: import { onAfterSignup } from "...",
       onBeforeOAuthRedirect: import { onBeforeOAuthRedirect } from "...",
-      onAfterOAuthTokenReceived: import { onAfterOAuthTokenReceived } from "...",
     },
   }
   ```